Abstracted out PremisLogger interface and changed the previous implementation to be named FilePremisLogger
