+++ fcrepo-clients/src/main/java/edu/unc/lib/dl/util/PremisLogger.java
@@ -15,118 +15,55 @@
  */
 package edu.unc.lib.dl.util;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.Date;
-import java.util.UUID;
+import java.util.List;
 
-import org.apache.jena.riot.RDFDataMgr;
-import org.apache.jena.riot.RDFFormat;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.hp.hpl.jena.rdf.model.Model;
-import com.hp.hpl.jena.rdf.model.ModelFactory;
 import com.hp.hpl.jena.rdf.model.Resource;
-import com.hp.hpl.jena.util.FileManager;
 
-import edu.unc.lib.dl.fcrepo4.RepositoryPathConstants;
+import edu.unc.lib.dl.fcrepo4.PremisEventObject;
+import edu.unc.lib.dl.fedora.FedoraException;
 import edu.unc.lib.dl.fedora.PID;
 
 /**
- * Logs PREMIS events for a repository object
  * 
- * @author lfarrell
+ * @author bbpennel
  *
  */
-public class PremisLogger {
-	private static final Logger log = LoggerFactory
-			.getLogger(PremisLogger.class);
-
-	private File premisFile;
-	private PID objectPid;
-	private Model model;
-
-	public PremisLogger(PID pid) {
-		this.objectPid = pid;
-	}
-
-	public PremisLogger(PID pid, File file) {
-		this.objectPid = pid;
-		this.premisFile = file;
-	}
+public interface PremisLogger {
 
 	/**
 	 * Allows for an arbitrary timestamp to be set for a premis event
 	 * @param eventType
 	 * @return PremisEventBuilder
 	 */
-	public PremisEventBuilder buildEvent(Resource eventType, Date date) {
-		if (date == null) {
-			date = new Date();
-		}
-
-		return new PremisEventBuilder(generateEventId(), eventType, date, this);
-	}
+	public PremisEventBuilder buildEvent(Resource eventType, Date date);
 
 	/**
 	 * Returns an instance of buildEvent with the timestamp automatically set to the current time
 	 * @param eventType
 	 * @return PremisEventBuilder
 	 */
-	public PremisEventBuilder buildEvent(Resource eventType) {
-		return new PremisEventBuilder(generateEventId(), eventType, new Date(), this);
-	}
+	public PremisEventBuilder buildEvent(Resource eventType);
 
 	/**
-	 * Adds an event to the log file
+	 * Adds an event to this log
 	 * 
 	 * @param eventResc
 	 * @return
 	 */
-	public PremisLogger writeEvent(Resource eventResc) {
-		// Add the event to the model for this event log
-		Model model = getModel().add(eventResc.getModel());
-
-		if (premisFile != null) {
-			// Persist the log to file
-			try (FileOutputStream rdfFile = new FileOutputStream(premisFile)) {
-				RDFDataMgr.write(rdfFile, model, RDFFormat.TURTLE_PRETTY);
-			} catch (IOException e) {
-				log.debug("Failed to serialize properties for object {} for the following reason {}", 
-					this.objectPid.getUUID(), e.getMessage());
-			}
-		}
-
-		return this;
-	}
+	public PremisLogger writeEvent(Resource eventResc);
 
 	/**
-	 * Returns the Model containing events from this logger
 	 * 
 	 * @return
+	 * @throws FedoraException
 	 */
-	public Model getModel() {
-		if (model != null) {
-			return model;
-		}
+	public List<PID> listEvents() throws FedoraException;
 
-		Model model = ModelFactory.createDefaultModel();
-
-		if (premisFile != null && premisFile.exists()) {
-			InputStream in = FileManager.get().open(premisFile.getAbsolutePath());
-			model.read(in, null, "TURTLE");
-		}
-
-		return model;
-	}
-
-	private String generateEventId() {
-		String uuid = UUID.randomUUID().toString();
-		String eventId = URIUtil.join(objectPid.getRepositoryPath(),
-				RepositoryPathConstants.EVENTS_CONTAINER, uuid);
-		return eventId;
-	}
+	/**
+	 * 
+	 * @return
+	 * @throws FedoraException
+	 */
+	public List<PremisEventObject> getEvents() throws FedoraException;
 }
